+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Methods with closure arguments can now be mocked.  Technically they always
   could be, but until now it wasn't possible to call the closure argument from
   `withf` or `returning`.  No special tricks are required by the user.
+  Similarly, methods with bare fn arguments can be mocked, too.
   ([#15](https://github.com/asomers/mockall/pull/15))
 
 ### Changed